chore: separate out feeding struct for clarity
This will eventually be combined with the main
app's Feeding struct, but put here now for better
clarity.
